PaperlessUiSettings data object
